feat: add get order by id endpoint

+++ cmd/local/order/application/usecases/create_order.go
@@ -19,7 +19,7 @@ type CreateOrderRequest struct {
 
 type CreateOrderItems struct {
 	ID        uuid.UUID
-	Quantity  int16
+	Quantity  int32
 	UnitPrice int64
 }
